requirements.txt
Flask==3.0.0
opencv-python==4.8.1.78
face_recognition==1.3.0
pandas==2.1.1
openpyxl==3.1.2
Werkzeug==3.0.0
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
python-dotenv==1.0.0
numpy<2.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
